# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ExperimentState": {
        "enum": [
          "CREATED",
          "VALIDATING",
          "BASELINING",
          "INJECTING",
          "OBSERVING",
          "RECOVERING",
          "COMPLETED",
          "VALIDATION_FAILED",
          "INJECTION_FAILED",
          "ABORTED",
          "UNKNOWN"
        ],
        "title": "ExperimentState",
        "type": "string"
      },
      "ExperimentSummary": {
        "description": "Slim experiment row for history lists and dashboards (no evidence JSON).",
        "properties": {
          "affected_replicas": {
            "title": "Affected Replicas",
            "type": "integer"
          },
          "auto": {
            "title": "Auto",
            "type": "boolean"
          },
          "client_outage_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Outage Seconds"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "integer"
          },
          "fault_type": {
            "$ref": "#/components/schemas/FaultType"
          },
          "has_baseline": {
            "title": "Has Baseline",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outcome_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Reason"
          },
          "pod_delete_mode": {
            "$ref": "#/components/schemas/PodDeleteMode"
          },
          "score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/ExperimentState"
          },
          "target": {
            "$ref": "#/components/schemas/ExperimentTargetSchema"
          },
          "time_to_recovery_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time To Recovery Seconds"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "target",
          "fault_type",
          "pod_delete_mode",
          "affected_replicas",
          "duration_seconds",
          "state",
          "auto",
          "has_baseline",
          "score",
          "time_to_recovery_seconds",
          "client_outage_seconds",
          "outcome_reason",
          "created_at",
          "updated_at"
        ],
        "title": "ExperimentSummary",
        "type": "object"
      },
      "ExperimentTargetSchema": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/WorkloadKind"
          },
          "name": {
            "maxLength": 253,
            "minLength": 1,
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$",
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "maxLength": 63,
            "minLength": 1,
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
            "title": "Namespace",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "kind",
          "name"
        ],
        "title": "ExperimentTargetSchema",
        "type": "object"
      },
      "FaultCoverage": {
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "fault_type": {
            "$ref": "#/components/schemas/FaultType"
          },
          "last_run_at": {
            "format": "date-time",
            "title": "Last Run At",
            "type": "string"
          },
          "latest_score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "pod_delete_mode": {
            "$ref": "#/components/schemas/PodDeleteMode"
          },
          "runs": {
            "title": "Runs",
            "type": "integer"
          }
        },
        "required": [
          "fault_type",
          "pod_delete_mode",
          "runs",
          "completed",
          "last_run_at",
          "latest_score"
        ],
        "title": "FaultCoverage",
        "type": "object"
      },
      "FaultType": {
        "enum": [
          "pod-delete",
          "pod-cpu-hog",
          "pod-memory-hog",
          "pod-network-latency"
        ],
        "title": "FaultType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LatestScore": {
        "description": "The most recent experiment on this workload that produced a numeric score.",
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "experiment_id": {
            "format": "uuid",
            "title": "Experiment Id",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "experiment_id",
          "score",
          "rating",
          "status",
          "version",
          "at"
        ],
        "title": "LatestScore",
        "type": "object"
      },
      "LiveStatus": {
        "description": "Pod-level status, read the same way safety validation reads it.",
        "properties": {
          "desired_replicas": {
            "title": "Desired Replicas",
            "type": "integer"
          },
          "ready_pod_names": {
            "items": {
              "type": "string"
            },
            "title": "Ready Pod Names",
            "type": "array"
          },
          "ready_pods": {
            "title": "Ready Pods",
            "type": "integer"
          },
          "running_pods": {
            "title": "Running Pods",
            "type": "integer"
          }
        },
        "required": [
          "desired_replicas",
          "running_pods",
          "ready_pods",
          "ready_pod_names"
        ],
        "title": "LiveStatus",
        "type": "object"
      },
...
      "PodDeleteMode": {
        "description": "How pod-delete removes pods (maps to the vetted experiment's FORCE env).\n\nGRACEFUL: default Kubernetes deletion; the pod's terminationGracePeriodSeconds\n          applies (SIGTERM, then SIGKILL after the grace period).\nFORCE:    deletion with gracePeriodSeconds=0; containers are killed at once.",
        "enum": [
          "GRACEFUL",
          "FORCE"
        ],
        "title": "PodDeleteMode",
        "type": "string"
      },
...
      "ScorePoint": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "experiment_id": {
            "format": "uuid",
            "title": "Experiment Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "workload": {
            "title": "Workload",
            "type": "string"
          }
        },
        "required": [
          "experiment_id",
          "name",
          "namespace",
          "workload",
          "at",
          "score",
          "rating"
        ],
        "title": "ScorePoint",
        "type": "object"
      },
...
      "ScoreSummary": {
        "properties": {
          "rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "score",
          "rating",
          "status",
          "version"
        ],
        "title": "ScoreSummary",
        "type": "object"
      },
      "ServiceDetail": {
        "properties": {
          "cluster_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Error"
          },
          "experiment_count": {
            "title": "Experiment Count",
            "type": "integer"
          },
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ExperimentSummary"
            },
            "title": "Experiments",
            "type": "array"
          },
          "faults": {
            "items": {
              "$ref": "#/components/schemas/FaultCoverage"
            },
            "title": "Faults",
            "type": "array"
          },
          "health": {
            "enum": [
              "HEALTHY",
              "DEGRADED",
              "UNAVAILABLE",
              "SCALED_TO_ZERO",
              "NOT_FOUND",
              "UNKNOWN"
            ],
            "title": "Health",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/WorkloadKind"
          },
          "latest_score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "live": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LiveStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "score_history": {
            "items": {
              "$ref": "#/components/schemas/ScorePoint"
            },
            "title": "Score History",
            "type": "array"
          },
          "score_version": {
            "title": "Score Version",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "kind",
          "name",
          "health",
          "live",
          "cluster_error",
          "experiment_count",
          "latest_score",
          "faults",
          "score_version",
          "score_history",
          "experiments"
        ],
        "title": "ServiceDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkloadKind": {
        "enum": [
          "Deployment",
          "StatefulSet"
        ],
        "title": "WorkloadKind",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "AutoResilience API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/services/{namespace}/{kind}/{name}": {
      "get": {
        "description": "Live pod health plus this workload's experiments, scores and tested faults.\n\nIf the cluster cannot be queried, history is still returned with\nhealth UNKNOWN and the error in `cluster_error`.",
        "operationId": "get_service_services__namespace___kind___name__get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "maxLength": 63,
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 253,
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Service",
        "tags": [
          "services"
        ]
      }
    }
  }
}